% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcSet.R
\name{arithmeticSet}
\alias{arithmeticSet}
\title{arithmeticSet}
\usage{
arithmeticSet()
}
\value{
the function set created
}
\description{
Creates a function set containing the basic arithmetic operators (+, -, *, /)
}
\examples{
arithmeticSet()
}
