% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{buildSolutionText}
\alias{buildSolutionText}
\title{buildSolutionText}
\usage{
buildSolutionText(functionNodes, functionSet, chromoID)
}
\arguments{
\item{functionNodes}{the structure of function nodes}

\item{functionSet}{the functionSet used when creating the functionNodes}

\item{chromoID}{the chromoID of the node to start at}
}
\value{
the character vector built
}
\description{
Used to recursively build up a character vector which shows the function
applied at the given node and each of its inputs.
}
