% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{calculateFitness}
\alias{calculateFitness}
\title{calculateFitness}
\usage{
calculateFitness(solution, dataModel, fitnessFunction, functionSet)
}
\arguments{
\item{solution}{the solution to be evaluated}

\item{dataModel}{a list containing the the dataset used to evaluate each
solution against and a formula for the model of the dataset}

\item{fitnessFunction}{the fitness function to use}

\item{functionSet}{the functionSet used with the population}
}
\value{
the fitness value of the solution
}
\description{
Calculates the fitness of a solution
}
