% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandler.R
\name{calculateInputSize}
\alias{calculateInputSize}
\title{calculateInputSize}
\usage{
calculateInputSize(model)
}
\arguments{
\item{model}{the formula to calculate the input size of}
}
\value{
the number of terms on the right hand side of the model
}
\description{
Calculates the input size from a given model. The input size is the
number of terms on the right hand side of the model.
}
