% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandler.R
\name{calculateOutputSize}
\alias{calculateOutputSize}
\title{calculateOutputSize}
\usage{
calculateOutputSize(model)
}
\arguments{
\item{model}{the formula to calculate the output size of}
}
\value{
the number of terms on the left hand side of the model
}
\description{
Calculates the output size from a given model. The output size is the
number of terms on the left hand side of the model.
}
\examples{
calculateOutputSize(output ~ x + y)
calculateOutputSize(Class + Type ~ x + y + z)
}
