% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{calculatePopFitness}
\alias{calculatePopFitness}
\title{calculatePopFitness}
\usage{
calculatePopFitness(population, dataset, fitnessFunction, functionSet)
}
\arguments{
\item{population}{the list of elements to be evaluated}

\item{dataset}{the dataset used to evaluate the element against}

\item{fitnessFunction}{the fitness function to use}

\item{functionSet}{the functionSet used with the population}
}
\value{
the population with fitness values nested inside
}
\description{
Calculates the fitness of each element in the population
}
