% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{calculateValue}
\alias{calculateValue}
\title{calculateValue}
\usage{
calculateValue(node, solution, functionSet)
}
\arguments{
\item{node}{the current functionNode}

\item{solution}{the solution containing the nodes}

\item{functionSet}{the functionSet used when creating the population}
}
\value{
the value for the currentNode
}
\description{
Calculates the output value after propagating the values of the inputNodes
through each of the functionNodes.
}
