% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{calculateValue2}
\alias{calculateValue2}
\title{calculateValue2}
\usage{
calculateValue2(node, solution, functionSet)
}
\arguments{
\item{node}{the current node}

\item{solution}{the solution containing all nodes}

\item{functionSet}{the functionSet used when creating the population}
}
\value{
the value calculated
}
\description{
Calculates the output value after propagating the values of the inputNodes
through each of the functionNodes. This function is similar to calculateValue
except that it is recursive.
}
