% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{calculateValueInSolution}
\alias{calculateValueInSolution}
\title{calculateValueInSolution}
\usage{
calculateValueInSolution(solution, functionNodesUsed, functionSet)
}
\arguments{
\item{solution}{the solution containing the nodes}

\item{functionNodesUsed}{the function nodes required to get an output value}

\item{functionSet}{the functionSet used when creating the population}
}
\value{
the solution after writing the calculated values in
}
\description{
Calculates the value that should be stored in each entry of functionNodesUsed
and stores it in the corresponding location of solution
}
