% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cgp}
\alias{cgp}
\title{cgp}
\usage{
cgp(dataset, model, functionSet = mathOpSet(), maxGenerations,
  fitnessFunction = mae, selectionMethod = list(func = muLambdaStrategy,
  args = c(population = NA, 4, NA)), rowsFuncNodes, colsFuncNodes, levelsBack,
  popSize = 5)
}
\arguments{
\item{dataset}{the dataset to use for training}

\item{model}{a model specifying the output and input fields of the dataset}

\item{functionSet}{the functions that can be used in the function nodes}

\item{maxGenerations}{the maximum generations of the evolutionary process}

\item{fitnessFunction}{the fitness function to use}

\item{selectionMethod}{the function and parameters to use for selection}

\item{rowsFuncNodes}{the number of rows to use in the function node structure}

\item{colsFuncNodes}{the number of columns to use in the function node structure}

\item{levelsBack}{the number of columns back that a function node can access}

\item{popSize}{the number of solutions to generate in each generation}
}
\value{
a list containing the best solution found and the chosen functionSet
}
\description{
Performs Cartesian Genetic Programming using the parameters provided.
}
