% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandler.R
\name{checkOutputBounds}
\alias{checkOutputBounds}
\title{checkOutputBounds}
\usage{
checkOutputBounds(numCol, outputColumns)
}
\arguments{
\item{numCol}{the number of columns in the dataset}

\item{outputColumns}{a vector containing integers to specify the columns}
}
\value{
a boolean of if the choices were valid or not
}
\description{
Checks that the columns chosen as outputs are valid choices. They are valid
if each column specified is within the range of 1 and numCol, and that at
least 1 column will remain to be used as input if using all output column
choices.
}
