% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{checkSolutionFound}
\alias{checkSolutionFound}
\title{checkSolutionFound}
\usage{
checkSolutionFound(population)
}
\arguments{
\item{population}{the population which contains fitness values}
}
\value{
a boolean stating whether a solution has been found or not
}
\description{
Checks if a solution has been found. A solution has been found if there is
a individual in the population with a fitness value of zero.
}
