% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcSet.R
\name{complexSet}
\alias{complexSet}
\title{complexSet}
\usage{
complexSet()
}
\value{
the function set created
}
\description{
Creates a function set containing the log, exp and sqrt functions
}
\examples{
complexSet()
}
