% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcSet.R
\name{constructFuncSet}
\alias{constructFuncSet}
\title{constructFuncSet}
\usage{
constructFuncSet(functionDefs)
}
\arguments{
\item{functionDefs}{the function definitions}
}
\value{
the function set created
}
\description{
Constructs a function set using the provided function definitions.
}
\examples{
constructFuncSet(c(c("add", 2), c("sqrt", 1)))
}
