% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{createFunctionNodesStructure}
\alias{createFunctionNodesStructure}
\title{createFunctionNodesStructure}
\usage{
createFunctionNodesStructure(rowsRequired)
}
\arguments{
\item{rowsRequired}{the number of rows to create}
}
\value{
the data frame created
}
\description{
Creates a data frame containing the structure of a function node. The data
frame is the length set in the generateFunctionNodes parameters.
}
