% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{decode}
\alias{decode}
\title{decode}
\usage{
decode(solution, functionNodesUsed, functionSet)
}
\arguments{
\item{solution}{the solution containing the nodes}

\item{functionNodesUsed}{the function nodes required to get an output value}

\item{functionSet}{the functionSet used when creating the population}
}
\value{
the output value
}
\description{
Decodes the solution and returns the output value of it.
}
