% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{decode2}
\alias{decode2}
\title{decode2}
\usage{
decode2(solution, functionSet)
}
\arguments{
\item{solution}{the solution containing the nodes}

\item{functionSet}{the functionSet used when creating the solution}
}
\value{
the value calculated
}
\description{
Decodes the solution. This should be used once the cgp function has passed
a solution out of the program as this decoding method produces a value
quicker
}
