% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandler.R
\name{extractRequiredFields}
\alias{extractRequiredFields}
\title{extractRequiredFields}
\usage{
extractRequiredFields(dataset, model)
}
\arguments{
\item{dataset}{the dataset to extract from}

\item{model}{a formula containing the variables to extract}
}
\value{
a dataset containing only the extracted fields
}
\description{
Extract the fields from dataset that are contained in model. It is a wrapper
for model.frame which provides a more informative handling of errors by
stating which variables were missing if any.
}
