% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{extractRequiredNodes}
\alias{extractRequiredNodes}
\title{extractRequiredNodes}
\usage{
extractRequiredNodes(solution)
}
\arguments{
\item{solution}{the solution containing the nodes to be extracted}
}
\value{
the required nodes
}
\description{
This function is to be used at the end of evolution so that only the required
nodes are passed out of the program.
}
