% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{findRow}
\alias{findRow}
\title{findRow}
\usage{
findRow(solution, chromoID)
}
\arguments{
\item{solution}{the solution containing a row with chromoID}

\item{chromoID}{the row to find}
}
\value{
the row found
}
\description{
Finds the row in the solution that contains chromoID
}
