% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getValidInputs}
\alias{getValidInputs}
\title{getValidInputs}
\usage{
getValidInputs(chromoID, functionNodeRange, functionNodeStructure)
}
\arguments{
\item{chromoID}{the chromoID to calculate the range of}

\item{functionNodeRange}{all the chromoIDs contained in functionNodes}

\item{functionNodeStructure}{the parameters used to create functionNodes}
}
\value{
the valid chromoIDs
}
\description{
Determines the valid range of input chromoIDs for a given chromoID
}
