% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{initPopulation}
\alias{initPopulation}
\title{initPopulation}
\usage{
initPopulation(popsize, functionSet, inputSize, outputSize, rowsFuncNodes,
  colsFuncNodes, levelsBack)
}
\arguments{
\item{popsize}{the number of solutions to be generate}

\item{functionSet}{the functions that can be used in the function nodes}

\item{inputSize}{the number of input nodes required. This should be the same as the number of input fields in the provided dataset}

\item{outputSize}{the number of output nodes required. This should be the same as the number of output fields in the provided dataset}

\item{rowsFuncNodes}{the number of rows to use in the function node structure}

\item{colsFuncNodes}{the number of columns to use in the function node structure}

\item{levelsBack}{the number of columns back that a function node can access}
}
\value{
a list containing the created population
}
\description{
Creates the initial population of solutions which is made up of popsize
lists. Each list contains three data frames, one for each type of node
(input, function, output).
}
