% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{is.functionNode}
\alias{is.functionNode}
\title{is.functionNode}
\usage{
is.functionNode(solution, chromoID)
}
\arguments{
\item{solution}{the solution containing the nodes}

\item{chromoID}{the chromoID of the node to check}
}
\value{
a logical value stating if the node is an functionNode or not
}
\description{
Checks if the node specified by chromoID is a functionNode
}
