% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{is.outputNode}
\alias{is.outputNode}
\title{is.outputNode}
\usage{
is.outputNode(solution, chromoID)
}
\arguments{
\item{solution}{the solution containing the nodes}

\item{chromoID}{the chromoID of the node to check}
}
\value{
a logical value stating if the node is an outputNode or not
}
\description{
Checks if the node specified by chromoID is an outputNode
}
