% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{mae}
\alias{mae}
\title{mae}
\usage{
mae(data)
}
\arguments{
\item{data}{a dataset containing predicted and actual values}
}
\value{
the mae value
}
\description{
Calculates the Mean Absolute Error on a set of predicted and actual values.
MAE produces a clear indicator of the average absolute difference between
predicted and actual values.
}
