% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{makeFunctionNode}
\alias{makeFunctionNode}
\title{makeFunctionNode}
\usage{
makeFunctionNode(chromoID)
}
\arguments{
\item{chromoID}{the unique ID of the node}
}
\value{
the node created
}
\description{
Creates a valid function node using the chromoID and value passed as a
parameter. A random function is chosen from the functionSet and random inputs
are chosen to satisfy the arity of the chosen function.
}
\examples{
makeFunctionNode(chromoID = 3)
makeFunctionNode(4)

}
