% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcSet.R
\name{mathOpSet}
\alias{mathOpSet}
\title{mathOpSet}
\usage{
mathOpSet()
}
\value{
the function set created
}
\description{
Creates a function set containing the functions from the other sets defined
in the package
}
\examples{
mathOpSet()
}
