% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{muLambdaStrategy}
\alias{muLambdaStrategy}
\title{muLambdaStrategy}
\usage{
muLambdaStrategy(population, lambda, functionNodeStructure)
}
\arguments{
\item{population}{the population to be evolved}

\item{lambda}{the number of offspring to generate}

\item{functionNodeStructure}{the parameters used when creating functionNodes}
}
\value{
a new population containing the parent used and the offspring
}
\description{
Runs the Mu + Lambda Evolutionary Strategy. Mu is set as 1 and Lambda can be
changed through the lambda parameter. Mu is the number of parents and Lambda
is the number of offspring to generate.
}
