% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{mutateFunction}
\alias{mutateFunction}
\title{mutateFunction}
\usage{
mutateFunction(solution, chromoID, functionNodeStructure)
}
\arguments{
\item{solution}{the solution containing the nodes}

\item{chromoID}{the node to be given a new function}

\item{functionNodeStructure}{the parameters used when creating functionNodes}
}
\value{
the solution after mutation
}
\description{
Generates a random new funcID for the node given by chromoID. Where the new
function has a different number of inputs than the current function, new
inputs are generated or removed as fit
}
