% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{nodesToProcess}
\alias{nodesToProcess}
\title{nodesToProcess}
\usage{
nodesToProcess(solution)
}
\arguments{
\item{solution}{The solution containing the nodes}
}
\value{
the functionNodes required
}
\description{
Find the functionNodes which are required by the outputNodes.
}
