% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plotGraph}
\alias{plotGraph}
\title{plotGraph}
\usage{
plotGraph(plotData)
}
\arguments{
\item{plotData}{the data to plot}
}
\description{
Plots the graph using ggplot2 and plotly and displays it in an interactive
shiny window through the browser.
}
