% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{printFinalDetails}
\alias{printFinalDetails}
\title{printFinalDetails}
\usage{
printFinalDetails(finalGeneration, maxGeneration, bestSolution, population)
}
\arguments{
\item{finalGeneration}{the final generation of evolution}

\item{maxGeneration}{the maximum generations for evolution}

\item{bestSolution}{the best solution found}

\item{population}{the population of all solutions}
}
\description{
Print information on the final result of evolution
}
