% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{printSolution}
\alias{printSolution}
\title{printSolution}
\usage{
printSolution(solution, functionSet)
}
\arguments{
\item{solution}{the solution to be printed as text}

\item{functionSet}{the functionSet used when creating the solution}
}
\value{
the textual format
}
\description{
Prints the solution in a textual format
}
