% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{rmse}
\alias{rmse}
\title{rmse}
\usage{
rmse(data)
}
\arguments{
\item{data}{a dataset containing predicted and actual values}
}
\value{
the rmse value
}
\description{
Calculates the Root Mean Squared Error on a set of predicted and actual
values. RMSE gives a large weight to large errors and should be used where
large errors are undesirable.
}
