% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{sampleWithoutBiasOrNA}
\alias{sampleWithoutBiasOrNA}
\title{sampleWithoutBiasOrNA}
\usage{
sampleWithoutBiasOrNA(x, size, replace = FALSE)
}
\arguments{
\item{x}{the vector to choose from}

\item{size}{the number of items to choose}

\item{replace}{should sampling be with replacement?}
}
\value{
the result of the sample function on x
}
\description{
Perform the sample function after removing any duplicate or NA values from
the vector
}
\examples{
sampleWithoutBiasOrNA(c(1, 2, 3, 1), 2)
sampleWithoutBiasOrNA(c(3, 4, NA), 4, replace = TRUE)
}
