% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{sortPopulation}
\alias{sortPopulation}
\title{sortPopulation}
\usage{
sortPopulation(population)
}
\arguments{
\item{population}{the population to be sorted}
}
\value{
the population after sorting
}
\description{
Sorts the population from lowest fitness value to highest
}
