% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{tournamentSelectionStrategy}
\alias{tournamentSelectionStrategy}
\title{tournamentSelectionStrategy}
\usage{
tournamentSelectionStrategy(population, tournamentSize, functionNodeStructure)
}
\arguments{
\item{population}{the population to be evolved}

\item{tournamentSize}{the number of samples to take}

\item{functionNodeStructure}{the parameters used when creating functionNodes}
}
\value{
a new population containing the parent used and the offspring
}
\description{
Runs Tournament Selection and returns a new population generated from the
solution Tournament Selection produced.
}
