% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandler.R
\name{transformDataset}
\alias{transformDataset}
\title{transformDataset}
\usage{
transformDataset(dataset, outputColumns)
}
\arguments{
\item{dataset}{the data frame to transform}

\item{outputColumns}{a vector containing integers to specify the columns}
}
\value{
the data frame after transformation
}
\description{
Transforms the given data frame so that the columns specified by outputColumns
are moved to the end of the data frame.
}
