% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcSet.R
\name{trigonometricSet}
\alias{trigonometricSet}
\title{trigonometricSet}
\usage{
trigonometricSet()
}
\value{
the function set created
}
\description{
Creates a function set containing the trigonometric operators (sin, cos, tan)
}
\examples{
trigonometricSet()
}
