% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{updateValidInputs}
\alias{updateValidInputs}
\title{updateValidInputs}
\usage{
updateValidInputs(row, level, validInputs)
}
\arguments{
\item{row}{the row to replace}

\item{level}{the new valid chromoIDs}

\item{validInputs}{a matrix containing the valid input chromoIDs}
}
\value{
a matrix containing the updated valid input chromoIDs
}
\description{
Replaces the no longer valid chromoIDs with the new level of
valid chromoIDs
}
