% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{generateInputs}
\alias{generateInputs}
\title{generateInputs}
\usage{
generateInputs()
}
\value{
a data frame containing the input nodes
}
\description{
Generates a data frame containing a row for each input attribute. Each row is
given a unique ID (chromoID) which is unique within the entire chromosome.
The value field of each row is set as NA initially.
}
