% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{generateOutputs}
\alias{generateOutputs}
\title{generateOutputs}
\usage{
generateOutputs(startID, maxInputID, outputSize)
}
\arguments{
\item{startID}{the starting chromoID}

\item{maxInputID}{the maximum chromoID that can be accessed}

\item{outputSize}{the number of output nodes required}
}
\value{
a data frame containing the output nodes
}
\description{
Generates a data frame containing a row for each output attribute. Each row
is given a unique ID (chromoID) which is unique within the entire chromosome.
The value field of each row is set as NA initially.
}
