% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{printEvolutionDetails}
\alias{printEvolutionDetails}
\title{printEvolutionDetails}
\usage{
printEvolutionDetails(currGeneration, maxGeneration, bestSolution, population,
  updateFreq, updateCount)
}
\arguments{
\item{currGeneration}{the current generation of evolution}

\item{maxGeneration}{the maximum generations for evolution}

\item{bestSolution}{the best solution found so far}

\item{population}{the population of all solutions}

\item{updateFreq}{how many generations pass between progress updates}

\item{updateCount}{how many generations have passed since last update}
}
\value{
updateCount + 1
}
\description{
Print information about the progress of evolution
}
