% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{printEvolutionDetails}
\alias{printEvolutionDetails}
\title{printEvolutionDetails}
\usage{
printEvolutionDetails(currGeneration, maxGeneration, bestSolution, population)
}
\arguments{
\item{currGeneration}{the current generation of evolution}

\item{maxGeneration}{the maximum generations for evolution}

\item{bestSolution}{the best solution found so far}

\item{population}{the population of all solutions}
}
\value{
returns NULL
}
\description{
Print information about the progress of evolution
}
