% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{updateValidInputs}
\alias{updateValidInputs}
\title{updateValidInputs}
\usage{
updateValidInputs(row, level)
}
\arguments{
\item{row}{the row to replace}

\item{level}{the new valid chromoIDs}
}
\description{
Replaces the no longer valid chromoIDs with the new level of
valid chromoIDs
}
\examples{
updateValidInputIDs(2, c(3, 4, 5))
updateValidInputIDs(4, c(7, 8))
}
