% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataHandler.R
\name{validSelectionInput}
\alias{validSelectionInput}
\title{validSelectionInput}
\usage{
validSelectionInput(arguments)
}
\arguments{
\item{arguments}{A list containing the function and arguments}
}
\value{
a boolean stating if the arguments are valid
}
\description{
Checks that the number of arguments provided matches the what the provided
function expects.
}
\examples{

foo <- function() {}

#Extra parameters to be added in the args member to match the parameters
#foo expects
validSelectionInput(list(func = foo, args = c(population = NA)))
}
